# openapi.yaml
openapi: 3.0.3
info:
  title: Donation Platform API
  version: 1.0.0
  description: |
    This API handles the core functionality of the Donation Management Platform.  
    It allows users to register, authenticate, and interact with donation posts.

    Donors can create and manage donation entries, while charities can browse,
    filter, and claim available donations. The system ensures proper tracking of
    quantities to prevent over-claiming and maintains clear communication between users.

    Authentication is handled using JSON Web Tokens (JWT), and access to certain
    endpoints is restricted based on user roles (donor, charity, admin).

    This documentation provides details on all available endpoints, request formats,
    and expected responses.
    TODO: Confirm exact success status codes where the documentation only says "Success response".
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
...
paths:
...
  /auth/profile:
    put:
      tags: [Authentication]
      summary: Update profile
      description: Updates name, phone, organization, or registrationNumber for the logged-in user.
      operationId: updateProfile
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateProfileRequest'}
            example: {name: Adem, phone: "0555111122"}
      responses:
        "200":
          description: Profile updated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
              example: {_id: USER_ID, name: Adem, email: adem@gmail.com, phone: "0555111122", role: donor, status: approved}
        "400": {$ref: '#/components/responses/BadRequest'}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /auth/charities:
    get:
      tags: [Admin]
      summary: Get charities
      description: Returns charity accounts. Admin only.
      operationId: getCharities
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/CharityStatusQuery'}
      responses:
        "200":
          description: Charity accounts returned.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/CharityUser'}}
              example: [{_id: CHARITY_USER_ID, name: Haithem, email: haithem@gmail.com, phone: "0555000003", role: charity, organization: KhirCharity, registrationNumber: KHR2026001, status: pending}]
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /auth/charities/{id}/status:
    put:
      tags: [Admin]
      summary: Update charity status
      description: Approves, rejects, or resets a charity account to pending. Approved or rejected charities receive a notification.
      operationId: updateCharityStatus
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateCharityStatusRequest'}
            example: {status: approved}
      responses:
        "200":
          description: Charity status updated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CharityUser'}
              example: {_id: CHARITY_USER_ID, name: Haithem, email: haithem@gmail.com, role: charity, organization: KhirCharity, registrationNumber: KHR2026001, status: approved}
        "400": {$ref: '#/components/responses/BadRequest'}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /posts:
    get:
      tags: [Donations]
      summary: Get available donations
      description: Returns available donations. Public route.
      operationId: getAvailableDonations
      parameters:
        - {$ref: '#/components/parameters/DonationTypeQuery'}
        - {$ref: '#/components/parameters/SearchQuery'}
        - {$ref: '#/components/parameters/DonationSortQuery'}
      responses:
        "200":
          description: Donations returned.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/DonationListItem'}}
              example: [{_id: DONATION_ID, donor: {_id: DONOR_ID, name: Adem, email: adem@gmail.com}, name: Rice bags, donationType: food, quantity: 20, remainingQty: 20, description: Clean rice bags ready for pickup., status: available, contactPhone: "0555000001", contactEmail: adem@gmail.com}]
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
    post:
      tags: [Donations]
      summary: Create donation
      description: Creates a donation. Donor only.
      operationId: createDonation
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateDonationRequest'}
            example: {name: School notebooks, donationType: school_supplies, quantity: 40, description: New notebooks for children., contactPhone: "0555000001", contactEmail: adem@gmail.com}
      responses:
        "201":
          description: "Donation created. TODO: Confirm exact status code."
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Donation'}
              example: {_id: DONATION_ID, donor: DONOR_ID, name: School notebooks, donationType: school_supplies, quantity: 40, remainingQty: 40, description: New notebooks for children., status: available, contactPhone: "0555000001", contactEmail: adem@gmail.com}
        "400": {$ref: '#/components/responses/BadRequest'}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /posts/my:
    get:
      tags: [Donations]
      summary: Get my donations
      description: Returns donations created by the logged-in donor.
      operationId: getMyDonations
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Donor donations returned.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Donation'}}
              example: [{_id: DONATION_ID, donor: DONOR_ID, name: Winter clothes, donationType: clothes, quantity: 12, remainingQty: 12, description: Jackets and sweaters in good condition., status: available}]
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
  /posts/{id}:
    get:
      tags: [Donations]
      summary: Get one donation
      operationId: getDonationById
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        "200":
          description: Donation returned.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DonationDetail'}
              example: {_id: DONATION_ID, donor: {_id: DONOR_ID, name: Adem, email: adem@gmail.com, phone: "0555000001"}, name: Rice bags, donationType: food, quantity: 20, remainingQty: 20, description: Clean rice bags ready for pickup., status: available, contactPhone: "0555000001", contactEmail: adem@gmail.com}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
    put:
      tags: [Donations]
      summary: Update donation
      description: Updates a donation created by the logged-in donor. Cannot be edited after any quantity has been claimed.
      operationId: updateDonation
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateDonationRequest'}
            example: {name: School notebooks and pens, quantity: 50, description: New notebooks and blue pens.}
      responses:
        "200":
          description: Donation updated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Donation'}
              example: {_id: DONATION_ID, donor: DONOR_ID, name: School notebooks and pens, donationType: school_supplies, quantity: 50, remainingQty: 50, description: New notebooks and blue pens., status: available}
        "400": {$ref: '#/components/responses/BadRequest'}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
    delete:
      tags: [Donations]
      summary: Delete donation
      description: Deletes a donation created by the logged-in donor. Cannot be deleted after claims have started.
      operationId: deleteDonation
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        "200":
          description: Donation deleted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MessageResponse'}
              example: {message: Donation deleted successfully}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "403": {$ref: '#/components/responses/Forbidden'}
        "404": {$ref: '#/components/responses/NotFound'}
        "429": {$ref: '#/components/responses/RateLimit'}
        "500": {$ref: '#/components/responses/ServerError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token returned by POST /auth/login.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource identifier.
      schema: {type: string}
      example: RESOURCE_ID
    CharityStatusQuery:
      name: status
      in: query
      required: false
      description: Filter charities by status.
      schema: {$ref: '#/components/schemas/CharityStatus'}
      example: pending
...
    DonationTypeQuery:
      name: type
      in: query
      required: false
      description: Filter by donation type.
      schema: {$ref: '#/components/schemas/DonationType'}
      example: food
    SearchQuery:
      name: search
      in: query
      required: false
      description: Search inside donation name and description.
      schema: {type: string}
      example: rice
    DonationSortQuery:
      name: sort
      in: query
      required: false
      description: Sort donations.
      schema: {type: string, enum: [newest, oldest, a-z, z-a, quantity]}
      example: quantity
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          examples:
            validation: {value: {message: Invalid request data}}
            invalidDonationType: {value: {message: Invalid donation type}}
            invalidStatus: {value: {message: Invalid claim status}}
    Unauthorized:
      description: Not authenticated.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          examples:
            noToken: {value: {message: No token}}
            invalidToken: {value: {message: Invalid token}}
            tokenExpired: {value: {message: Token expired}}
            userNotFound: {value: {message: User not found}}
    Forbidden:
      description: Not authorized.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          examples:
            actionNotAuthorized: {value: {message: Action not authorized}}
            accessDenied: {value: {message: Access denied}}
            charityNotVerified: {value: {message: Charity not verified yet}}
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {message: Resource not found}
...
    RateLimit:
      description: Too many requests.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {message: Too many requests, please try again later}
...
    ServerError:
      description: Server error.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {message: Server error message}
...
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message: {type: string, example: Error message here}
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: {type: string, example: Operation completed successfully}
    UserRole:
      type: string
      enum: [admin, donor, charity]
    CharityStatus:
      type: string
      enum: [pending, approved, rejected]
    UserStatus:
      type: string
      enum: [pending, approved, rejected]
    DonationType:
      type: string
      enum: [food, clothes, toys, electronics, books, medicines, furniture, school_supplies, hygiene, baby_items, household, other, others]
    DonationStatus:
      type: string
      description: "TODO: Documentation shows available and completed; confirm the full backend enum."
      enum: [available, completed]
...
    UpdateProfileRequest:
      type: object
      description: At least one allowed profile field should be provided.
      properties:
        name: {type: string, example: Adem}
        phone: {type: string, example: "0555111122"}
        organization: {type: string, example: KhirCharity}
        registrationNumber: {type: string, example: KHR2026001}
    User:
      type: object
      required: [_id, name, email, role, status]
      properties:
        _id: {type: string, example: USER_ID}
        name: {type: string, example: Adem}
        email: {type: string, format: email, example: adem@gmail.com}
        phone: {type: string, example: "0555000001"}
        role: {$ref: '#/components/schemas/UserRole'}
        organization: {type: string, example: KhirCharity}
        registrationNumber: {type: string, example: KHR2026001}
        status: {$ref: '#/components/schemas/UserStatus'}
        createdAt: {type: string, format: date-time, example: "2026-04-12T20:00:00.000Z"}
        updatedAt: {type: string, format: date-time, example: "2026-04-12T20:00:00.000Z"}
    CharityUser:
      allOf:
        - {$ref: '#/components/schemas/User'}
        - type: object
          required: [role, organization, registrationNumber]
          properties:
            role: {type: string, enum: [charity]}
            organization: {type: string, example: KhirCharity}
            registrationNumber: {type: string, example: KHR2026001}
            status: {$ref: '#/components/schemas/CharityStatus'}
    UpdateCharityStatusRequest:
      type: object
      required: [status]
      properties:
        status: {$ref: '#/components/schemas/CharityStatus'}
    DonorSummary:
      type: object
      required: [_id, name, email]
      properties:
        _id: {type: string, example: DONOR_ID}
        name: {type: string, example: Adem}
        email: {type: string, format: email, example: adem@gmail.com}
        phone: {type: string, example: "0555000001"}
...
    DonationBase:
      type: object
      required: [_id, name, donationType, quantity, remainingQty, status]
      properties:
        _id: {type: string, example: DONATION_ID}
        name: {type: string, example: Rice bags}
        donationType: {$ref: '#/components/schemas/DonationType'}
        quantity: {type: integer, minimum: 1, example: 20}
        remainingQty: {type: integer, minimum: 0, example: 20}
        description: {type: string, example: Clean rice bags ready for pickup.}
        status: {$ref: '#/components/schemas/DonationStatus'}
        contactPhone: {type: string, example: "0555000001"}
        contactEmail: {type: string, format: email, example: adem@gmail.com}
    Donation:
      allOf:
        - {$ref: '#/components/schemas/DonationBase'}
        - type: object
          required: [donor]
          properties:
            donor: {type: string, example: DONOR_ID}
    DonationListItem:
      allOf:
        - {$ref: '#/components/schemas/DonationBase'}
        - type: object
          required: [donor]
          properties:
            donor: {$ref: '#/components/schemas/DonorSummary'}
    DonationDetail:
      allOf:
        - {$ref: '#/components/schemas/DonationBase'}
        - type: object
          required: [donor]
          properties:
            donor: {$ref: '#/components/schemas/DonorSummary'}
    CreateDonationRequest:
      type: object
      required: [name, donationType, quantity]
      properties:
        name: {type: string, example: School notebooks}
        donationType: {$ref: '#/components/schemas/DonationType'}
        quantity: {type: integer, minimum: 1, example: 40}
        description: {type: string, example: New notebooks for children.}
        contactPhone: {type: string, example: "0555000001"}
        contactEmail: {type: string, format: email, example: adem@gmail.com}
    UpdateDonationRequest:
      type: object
      description: At least one valid donation field should be provided.
      properties:
        name: {type: string, example: School notebooks and pens}
        donationType: {$ref: '#/components/schemas/DonationType'}
        quantity: {type: integer, minimum: 1, example: 50}
        description: {type: string, example: New notebooks and blue pens.}
        contactPhone: {type: string, example: "0555000001"}
        contactEmail: {type: string, format: email, example: adem@gmail.com}
...